+++ Src/Private/Export-Diagrammer.ps1
@@ -46,7 +46,7 @@ function Export-Diagrammer {
     .NOTES
         Version:        0.2.31
         Author:         Jonathan Colon
-        Twitter:        @jcolonfzenpr
+        Bluesky:        @jcolonfpr.bsky.social
         GitHub:         rebelinux
 
     .LINK